+++ src/main/java/dev/ftb/mods/ftbteambases/data/bases/BaseInstanceManager.java
@@ -79,7 +79,8 @@ public class BaseInstanceManager extends SavedData {
             BlockPos.CODEC.fieldOf("lobby_spawn_pos").forGetter(mgr -> mgr.lobbySpawnPos),
             NETHER_PORTAL_POS_CODEC.fieldOf("nether_portal_pos").forGetter(mgr -> mgr.playerNetherPortalLocs),
             PLAYER_ID_LIST_CODEC.fieldOf("known_players").forGetter(mgr -> mgr.knownPlayers),
-            PLAYER_ID_LIST_CODEC.optionalFieldOf("orphaned_players", new HashSet<>()).forGetter(mgr -> mgr.orphanedPlayers)
+            PLAYER_ID_LIST_CODEC.optionalFieldOf("orphaned_players", new HashSet<>()).forGetter(mgr -> mgr.orphanedPlayers),
+            Codec.BOOL.optionalFieldOf("autoclaim_needed", true).forGetter(mgr -> mgr.autoclaimNeeded)
     ).apply(inst, BaseInstanceManager::new));
 
     // maps team UUID to live base details
@@ -100,11 +101,13 @@ public class BaseInstanceManager extends SavedData {
     private boolean isLobbyCreated;
     private @Nullable BlockPos lobbySpawnPos;
     private int nextArchiveId;
+    private boolean autoclaimNeeded;
 
     private BaseInstanceManager(Map<UUID, LiveBaseDetails> liveBases, Map<ResourceLocation, RegionCoords> genPos,
                                 Map<ResourceLocation, Integer> zOffsets, Map<String, ArchivedBaseDetails> archivedBases,
                                 int nextArchiveId, boolean isLobbyCreated, @Nullable BlockPos lobbySpawnPos,
-                                Map<UUID,BlockPos> netherPortalPos, Set<UUID> knownPlayers, Set<UUID> orphanedPlayers) {
+                                Map<UUID,BlockPos> netherPortalPos, Set<UUID> knownPlayers, Set<UUID> orphanedPlayers,
+                                boolean autoclaimNeeded) {
         this.liveBases = liveBases;
         this.storedGenPos = genPos;
         this.storedZoffset = zOffsets;
@@ -119,7 +122,8 @@ public class BaseInstanceManager extends SavedData {
 
     private static BaseInstanceManager createNew() {
         return new BaseInstanceManager(new HashMap<>(), new HashMap<>(), new HashMap<>(), new HashMap<>(),
-                0, false, null, new HashMap<>(), new HashSet<>(), new HashSet<>());
+                0, false, null, new HashMap<>(), new HashSet<>(), new HashSet<>(),
+                true);
     }
 
     public static BaseInstanceManager get() {
@@ -451,6 +455,15 @@ public class BaseInstanceManager extends SavedData {
 
     public boolean isPlayerKnown(ServerPlayer player) {
         return knownPlayers.contains(player.getUUID());
+    }
+
+    public boolean isAutoclaimNeeded() {
+        return autoclaimNeeded;
+    }
+
+    public void setAutoclaimNeeded(boolean autoclaimNeeded) {
+        this.autoclaimNeeded = autoclaimNeeded;
+        setDirty();
     }
 
     public void forceSave(MinecraftServer server) {